# api/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Observable Go Backend Starter Kit API
  version: 1.1.0
  description: Cookie-based auth with Google OAuth2 and local email/password auth, plus RBAC-protected admin endpoints. Error responses support envelope JSON by default and RFC7807 via Accept negotiation (`application/problem+json`).
servers:
  - url: http://localhost:8080/api/v1
...
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Unique request key used to deduplicate retried mutating requests.
      schema:
        type: string
        minLength: 1
        maxLength: 128
      example: 8f08db4b-3173-42f8-9bc2-c97d2229b3cb
  schemas:
...
    Envelope:
      type: object
      properties:
        success: { type: boolean }
        data: {}
        error: { $ref: '#/components/schemas/APIError' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    UserListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/UserListData'
        meta:
          $ref: '#/components/schemas/Meta'
...
  responses:
    BadRequestError:
      description: Request payload/path/query is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalidPayload:
              value:
                success: false
                error:
                  code: BAD_REQUEST
                  message: invalid payload
                meta:
                  request_id: req-abc123
                  timestamp: "2026-02-09T10:00:00Z"
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ConflictError:
      description: Request conflicts with existing state (for example, idempotency key replay mismatch).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            conflict:
              value:
                success: false
                error:
                  code: CONFLICT
                  message: idempotency key reuse with different payload
                meta:
                  request_id: req-abc123
                  timestamp: "2026-02-09T10:00:00Z"
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    UnauthorizedError:
      description: Authentication failed or access token is missing/invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            unauthorized:
              value:
                success: false
                error:
                  code: UNAUTHORIZED
                  message: invalid credentials
                meta:
                  request_id: req-abc123
                  timestamp: "2026-02-09T10:00:00Z"
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
paths:
...
  /auth/google/callback:
    get:
      tags: [Auth]
      summary: Handle Google OAuth callback
      operationId: authGoogleCallback
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: state
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/local/register:
    post:
      tags: [Auth]
      summary: Register local email/password account
      operationId: authLocalRegister
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email: { type: string, format: email }
                name: { type: string }
                password: { type: string, format: password, minLength: 12 }
      responses:
        '201':
          description: Local registration success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /auth/local/login:
    post:
      tags: [Auth]
      summary: Login with local email/password
      operationId: authLocalLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Local login success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /auth/local/verify/request:
    post:
      tags: [Auth]
      summary: Request local email verification token
      operationId: authLocalVerifyRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202':
          description: Verification request accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/local/verify/confirm:
    post:
      tags: [Auth]
      summary: Confirm local email verification token
      operationId: authLocalVerifyConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/local/password/forgot:
    post:
      tags: [Auth]
      summary: Request local password reset token
      operationId: authLocalPasswordForgot
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Request accepted with generic response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/local/password/reset:
    post:
      tags: [Auth]
      summary: Reset local password using one-time token
      operationId: authLocalPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string }
                new_password: { type: string, format: password, minLength: 12 }
      responses:
        '200':
          description: Password reset success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/local/change-password:
    post:
      tags: [Auth]
      summary: Change local account password
      operationId: authLocalChangePassword
      security:
        - accessTokenCookie: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 12 }
      responses:
        '200': { description: Password changed }
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new access token
      operationId: authRefresh
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Token refresh success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke current user sessions
      operationId: authLogout
      security:
        - accessTokenCookie: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      responses:
        '200': { description: Logout success }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me:
    get:
      tags: [User]
      summary: Get current profile
      operationId: userMe
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me/sessions:
    get:
      tags: [User]
      summary: List active sessions for current user
      operationId: userListSessions
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Active sessions returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me/sessions/{session_id}:
    delete:
      tags: [User]
      summary: Revoke a specific session/device
      operationId: userRevokeSession
      security:
        - accessTokenCookie: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 1
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /me/sessions/revoke-others:
    post:
      tags: [User]
      summary: Revoke all active sessions except current one
      operationId: userRevokeOtherSessions
      security:
        - accessTokenCookie: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Other sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeOtherSessionsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /admin/users:
    get:
      tags: [Admin]
      summary: List users
      description: Returns paginated users with optional filtering by email/status/role and validated sorting.
      operationId: adminListUsers
      security:
        - accessTokenCookie: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: sort_by
          schema:
            type: string
            enum: [id, created_at, updated_at, email, name, status, last_login_at]
            default: created_at
        - in: query
          name: sort_order
          schema: { type: string, enum: [asc, desc], default: desc }
        - in: query
          name: email
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: role
          schema: { type: string }
      responses:
        '200':
          description: Users returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
              examples:
                users:
                  value:
                    success: true
                    data:
                      items:
                        - id: 42
                          email: admin@example.com
                          name: Admin User
                          avatar_url: https://cdn.example.com/avatars/42.png
                          status: active
                          last_login_at: "2026-02-09T09:58:32Z"
                          created_at: "2026-02-01T00:00:00Z"
                          updated_at: "2026-02-09T09:58:32Z"
                          roles:
                            - id: 1
                              name: admin
                              description: Platform administrators
                              created_at: "2026-02-01T00:00:00Z"
                              updated_at: "2026-02-01T00:00:00Z"
                      pagination:
                        page: 1
                        page_size: 20
                        total: 1
                        total_pages: 1
                    meta:
                      request_id: req-abc123
                      timestamp: "2026-02-09T10:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/users/{id}/roles:
    patch:
      tags: [Admin]
      summary: Set roles for user
      description: Replaces a user's assigned roles with the provided role ID set.
      operationId: adminSetUserRoles
      security:
        - accessTokenCookie: []
      parameters:
        - in: path
          name: id
          required: true
          description: Numeric user ID.
          schema:
            type: integer
            format: uint64
            minimum: 1
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserRolesRequest'
            examples:
              updateUserRoles:
                value:
                  role_ids: [1, 2]
      responses:
        '200':
          description: Roles updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetUserRolesResponse'
              examples:
                updated:
                  value:
                    success: true
                    data:
                      user_id: 42
                      role_ids: [1, 2]
                    meta:
                      request_id: req-abc123
                      timestamp: "2026-02-09T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/roles:
    get:
      tags: [Admin]
      summary: List roles
      description: Returns paginated RBAC roles with optional name filter and validated sorting.
      operationId: adminListRoles
      security:
        - accessTokenCookie: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: sort_by
          schema:
            type: string
            enum: [id, created_at, updated_at, name]
            default: created_at
        - in: query
          name: sort_order
          schema: { type: string, enum: [asc, desc], default: desc }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200':
          description: Roles returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
              examples:
                roles:
                  value:
                    success: true
                    data:
                      items:
                        - id: 1
                          name: admin
                          description: Platform administrators
                          permissions:
                            - id: 1
                              resource: users
                              action: read
                              created_at: "2026-02-01T00:00:00Z"
                              updated_at: "2026-02-01T00:00:00Z"
                            - id: 2
                              resource: users
                              action: write
                              created_at: "2026-02-01T00:00:00Z"
                              updated_at: "2026-02-01T00:00:00Z"
                          created_at: "2026-02-01T00:00:00Z"
                          updated_at: "2026-02-01T00:00:00Z"
                      pagination:
                        page: 1
                        page_size: 20
                        total: 2
                        total_pages: 1
                    meta:
                      request_id: req-abc123
                      timestamp: "2026-02-09T10:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Admin]
      summary: Create role
      description: Creates a new role and optionally attaches permissions by `resource:action` strings.
      operationId: adminCreateRole
      security:
        - accessTokenCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
            examples:
              createRole:
                value:
                  name: support_admin
                  description: Support operators with user management capabilities
                  permissions: [users:read, users:write]
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRoleResponse'
              examples:
                created:
                  value:
                    success: true
                    data:
                      id: 7
                      name: support_admin
                      description: Support operators with user management capabilities
                      permissions: []
                      created_at: "2026-02-09T10:00:00Z"
                      updated_at: "2026-02-09T10:00:00Z"
                    meta:
                      request_id: req-abc123
                      timestamp: "2026-02-09T10:00:00Z"
        '400':
          description: Invalid input (including malformed permission format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                invalidPermissionFormat:
                  value:
                    success: false
                    error:
                      code: BAD_REQUEST
                      message: permission format must be resource:action
                    meta:
                      request_id: req-abc123
                      timestamp: "2026-02-09T10:00:00Z"
                duplicateRole:
                  value:
                    success: false
                    error:
                      code: BAD_REQUEST
                      message: failed to create role
                    meta:
                      request_id: req-abc123
                      timestamp: "2026-02-09T10:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/permissions:
    get:
      tags: [Admin]
      summary: List permissions
      description: Returns paginated permission catalog with optional resource/action filters and validated sorting.
      operationId: adminListPermissions
      security:
        - accessTokenCookie: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: sort_by
          schema:
            type: string
            enum: [id, created_at, updated_at, resource, action]
            default: created_at
        - in: query
          name: sort_order
          schema: { type: string, enum: [asc, desc], default: desc }
        - in: query
          name: resource
          schema: { type: string }
        - in: query
          name: action
          schema: { type: string }
      responses:
        '200':
          description: Permissions returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
              examples:
                permissions:
                  value:
                    success: true
                    data:
                      items:
                        - id: 1
                          resource: users
                          action: read
                          created_at: "2026-02-01T00:00:00Z"
                          updated_at: "2026-02-01T00:00:00Z"
                        - id: 2
                          resource: users
                          action: write
                          created_at: "2026-02-01T00:00:00Z"
                          updated_at: "2026-02-01T00:00:00Z"
                      pagination:
                        page: 1
                        page_size: 20
                        total: 6
                        total_pages: 1
                    meta:
                      request_id: req-abc123
                      timestamp: "2026-02-09T10:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Admin]
      summary: Create permission
      description: Creates a permission using strict `resource:action` parts.
      operationId: adminCreatePermission
      security:
        - accessTokenCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermissionRequest'
      responses:
        '201':
          description: Permission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Permission conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/roles/{id}:
    patch:
      tags: [Admin]
      summary: Update role
      description: Updates role name/description and replaces permission bindings.
      operationId: adminUpdateRole
      security:
        - accessTokenCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRoleResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Role conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Admin]
      summary: Delete role
      description: Deletes a role unless protected by server policy.
      operationId: adminDeleteRole
      security:
        - accessTokenCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 1
      responses:
        '200':
          description: Role deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteRoleResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/rbac/sync:
    post:
      tags: [Admin]
      summary: Reconcile RBAC seed definitions
      description: Idempotently ensures default roles/permissions and bindings are present.
      operationId: adminSyncRBAC
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Sync report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RBACSyncResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'